Shift matching to relevant card companies
The validator shouldn't really know what methods the
card companies use to validate whether a card belongs to
them or not.

Each company implements a '.belongs_to_me?' method which
can be accessed by the validator. The checks for initial digits
and card length are then done within the respective company class